Moved SpotifyAuthResponse from network to spotify package
