+++ code/app/src/main/java/com/example/impostersyndrom/spotify/SpotifyAuthResponse.java
@@ -1,4 +1,4 @@
-package com.example.impostersyndrom.network;
+package com.example.impostersyndrom.spotify;
 
 public class SpotifyAuthResponse {
     public String access_token;
